Separate RealTime, Storage and Firestore from the DatabaseConnection
