+++ app/src/main/java/com/github/se/studybuddies/viewModels/DirectMessageViewModel.kt
@@ -3,7 +3,7 @@ package com.github.se.studybuddies.viewModels
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.se.studybuddies.data.Chat
-import com.github.se.studybuddies.database.DatabaseConnection
+import com.github.se.studybuddies.database.RealtimeDatabaseConnection
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 
 class DirectMessageViewModel(userUid: String = "") : ViewModel() {
-  private val db = DatabaseConnection()
+  private val db = RealtimeDatabaseConnection()
   private val _userUid = MutableStateFlow(userUid)
   private val _directMessages = MutableStateFlow<List<Chat>>(emptyList())
   val directMessages =
